01 - BE - DB - Initial implementation of transfer logic

+++ 01-basic-web-app/backend/repository/src/main/java/org/digitalcrafting/eregold/repository/transactions/TransactionsMapper.java
@@ -3,14 +3,11 @@ package org.digitalcrafting.eregold.repository.transactions;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 
-import java.math.BigInteger;
-
-// TODO introduce EntityManager
 @Mapper
 public interface TransactionsMapper {
 
     @Select("SELECT nextval('transaction_id_seq')")
-    BigInteger getNextId();
+    Long getNextId();
 
     void insert(TransactionEntity entity);
 }
